Refactor Reporting DTOs and Schema to use Integer Pennies
Harden Reporting DTOs, Watchtower DTOs, and Database Schema to strictly use `int` (pennies) for all monetary fields, eliminating floating-point drift in analytics and persistence.

Key Changes:
- `AgentTickAnalyticsDTO`: `labor_income` and `capital_income` are now `int`.
- `WatchtowerSnapshotDTO`: All nested monetary fields (GDP, M2, etc.) are `int`.
- `AgentStateData`: `assets` is now `Dict[CurrencyCode, int]`.
- `TransactionData`: Added `total_pennies: int` as SSoT.
- `EconomicIndicatorData`: Monetary aggregates are `int`.
- `schema.py`: Updated columns to `INTEGER` and added `total_pennies` to `transactions`.
- `AnalyticsSystem`: Updated aggregation logic to use integer math.
- Updated tests (`test_persistence_purity.py`, `test_watchtower_hardening.py`, `test_repository.py`) to reflect these changes.

+++ modules/analytics/dtos.py
@@ -8,8 +8,8 @@ class AgentTickAnalyticsDTO(TypedDict):
     run_id: int
     time: int
     agent_id: int
-    labor_income_this_tick: Optional[float]
-    capital_income_this_tick: Optional[float]
+    labor_income_this_tick: Optional[int]
+    capital_income_this_tick: Optional[int]
     consumption_this_tick: Optional[float]
     utility_this_tick: Optional[float]
     savings_rate_this_tick: Optional[float]
